rest controller, not tested

+++ src/main/java/be/kdg/outfitly/repository/ArduinoSensorRepository.java
@@ -3,5 +3,5 @@ package be.kdg.outfitly.repository;
 import be.kdg.outfitly.domain.ArduinoSensor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArduinoSensorRepository extends JpaRepository<ArduinoSensor, Integer> {
+public interface ArduinoSensorRepository extends JpaRepository<ArduinoSensor, Long> {
 }
